# Conflicts:
#	MANIFEST.in
#	PPOCRLabel/PPOCRLabel.py
#	PPOCRLabel/README.md
#	PPOCRLabel/README_ch.md
#	PPOCRLabel/libs/autoDialog.py
#	PPOCRLabel/requirements.txt
#	PPOCRLabel/setup.py
#	README.md
#	README_ch.md
#	applications/README.md
#	deploy/paddlejs/README.md
#	deploy/paddlejs/README_ch.md
#	deploy/slim/quantization/README.md
#	deploy/slim/quantization/export_model.py
#	doc/doc_ch/PP-OCRv3_det_train.md
#	doc/doc_ch/algorithm_det_east.md
#	doc/doc_ch/finetune.md
#	doc/doc_ch/quickstart.md
#	doc/doc_en/finetune_en.md
#	doc/doc_en/models_list_en.md
#	doc/doc_en/ocr_book_en.md
#	doc/doc_en/quickstart_en.md
#	paddleocr.py
#	requirements.txt
#	setup.py
#	train.sh
#	unstructured_paddleocr/paddle_tools/infer/predict_system.py
#	unstructured_paddleocr/paddle_tools/infer/utility.py
#	unstructured_paddleocr/ppocr/data/imaug/__init__.py
#	unstructured_paddleocr/ppocr/data/imaug/abinet_aug.py
#	unstructured_paddleocr/ppocr/data/imaug/label_ops.py
#	unstructured_paddleocr/ppocr/data/imaug/rec_img_aug.py
#	unstructured_paddleocr/ppocr/data/multi_scale_sampler.py
#	unstructured_paddleocr/ppocr/losses/rec_aster_loss.py
#	unstructured_paddleocr/ppocr/losses/rec_cppd_loss.py
#	unstructured_paddleocr/ppocr/losses/rec_nrtr_loss.py
#	unstructured_paddleocr/ppocr/losses/rec_parseq_loss.py
#	unstructured_paddleocr/ppocr/losses/rec_satrn_loss.py
#	unstructured_paddleocr/ppocr/modeling/backbones/det_pp_lcnet_v2.py
#	unstructured_paddleocr/ppocr/modeling/backbones/rec_hgnet.py
#	unstructured_paddleocr/ppocr/modeling/backbones/rec_lcnetv3.py
#	unstructured_paddleocr/ppocr/modeling/backbones/rec_repvit.py
#	unstructured_paddleocr/ppocr/modeling/backbones/rec_shallow_cnn.py
#	unstructured_paddleocr/ppocr/modeling/backbones/rec_svtrnet.py
#	unstructured_paddleocr/ppocr/modeling/backbones/rec_svtrv2.py
#	unstructured_paddleocr/ppocr/modeling/backbones/rec_vit.py
#	unstructured_paddleocr/ppocr/modeling/backbones/rec_vit_parseq.py
#	unstructured_paddleocr/ppocr/modeling/heads/rec_cppd_head.py
#	unstructured_paddleocr/ppocr/modeling/heads/rec_nrtr_head.py
#	unstructured_paddleocr/ppocr/modeling/heads/rec_parseq_head.py
#	unstructured_paddleocr/ppocr/modeling/heads/rec_satrn_head.py
#	unstructured_paddleocr/ppocr/modeling/necks/intracl.py
#	unstructured_paddleocr/ppocr/postprocess/db_postprocess.py
#	unstructured_paddleocr/ppocr/utils/dict/README.md
#	unstructured_paddleocr/ppocr/utils/dict/bengali_dict.txt
#	unstructured_paddleocr/ppocr/utils/dict/bm_dict.txt
#	unstructured_paddleocr/ppocr/utils/dict/gujarati_dict.txt
#	unstructured_paddleocr/ppocr/utils/dict/kazakh_dict.txt
#	unstructured_paddleocr/ppocr/utils/dict/parseq_dict.txt
#	unstructured_paddleocr/ppocr/utils/loggers/vdl_logger.py
#	unstructured_paddleocr/ppocr/utils/utility.py
#	unstructured_paddleocr/ppstructure/kie/README.md
#	unstructured_paddleocr/ppstructure/pdf2word/pdf2word.py
#	unstructured_paddleocr/ppstructure/predict_system.py
#	unstructured_paddleocr/ppstructure/recovery/recovery_to_doc.py
#	unstructured_paddleocr/ppstructure/return_word_pos.md
#	unstructured_paddleocr/ppstructure/utility.py

+++ unstructured_paddleocr/ppocr/utils/network.py
@@ -20,16 +20,20 @@ from tqdm import tqdm
 
 from ppocr.utils.logging import get_logger
 
+MODELS_DIR = os.path.expanduser("~/.paddleocr/models/")
+
 
 def download_with_progressbar(url, save_path):
     logger = get_logger()
+    if save_path and os.path.exists(save_path):
+        logger.info(f"Path {save_path} already exists. Skipping...")
+        return
     response = requests.get(url, stream=True)
     if response.status_code == 200:
-        total_size_in_bytes = int(response.headers.get('content-length', 1))
+        total_size_in_bytes = int(response.headers.get("content-length", 1))
         block_size = 1024  # 1 Kibibyte
-        progress_bar = tqdm(
-            total=total_size_in_bytes, unit='iB', unit_scale=True)
-        with open(save_path, 'wb') as file:
+        progress_bar = tqdm(total=total_size_in_bytes, unit="iB", unit_scale=True)
+        with open(save_path, "wb") as file:
             for data in response.iter_content(block_size):
                 progress_bar.update(len(data))
                 file.write(data)
@@ -41,34 +45,43 @@ def download_with_progressbar(url, save_path):
 
 def maybe_download(model_storage_directory, url):
     # using custom model
-    tar_file_name_list = ['.pdiparams', '.pdiparams.info', '.pdmodel']
+    tar_file_name_list = [".pdiparams", ".pdiparams.info", ".pdmodel"]
     if not os.path.exists(
-            os.path.join(model_storage_directory, 'inference.pdiparams')
-    ) or not os.path.exists(
-            os.path.join(model_storage_directory, 'inference.pdmodel')):
-        assert url.endswith('.tar'), 'Only supports tar compressed package'
-        tmp_path = os.path.join(model_storage_directory, url.split('/')[-1])
-        print('download {} to {}'.format(url, tmp_path))
+        os.path.join(model_storage_directory, "inference.pdiparams")
+    ) or not os.path.exists(os.path.join(model_storage_directory, "inference.pdmodel")):
+        assert url.endswith(".tar"), "Only supports tar compressed package"
+        tmp_path = os.path.join(model_storage_directory, url.split("/")[-1])
+        print("download {} to {}".format(url, tmp_path))
         os.makedirs(model_storage_directory, exist_ok=True)
         download_with_progressbar(url, tmp_path)
-        with tarfile.open(tmp_path, 'r') as tarObj:
+        with tarfile.open(tmp_path, "r") as tarObj:
             for member in tarObj.getmembers():
                 filename = None
                 for tar_file_name in tar_file_name_list:
                     if member.name.endswith(tar_file_name):
-                        filename = 'inference' + tar_file_name
+                        filename = "inference" + tar_file_name
                 if filename is None:
                     continue
                 file = tarObj.extractfile(member)
-                with open(
-                        os.path.join(model_storage_directory, filename),
-                        'wb') as f:
+                with open(os.path.join(model_storage_directory, filename), "wb") as f:
                     f.write(file.read())
         os.remove(tmp_path)
 
 
+def maybe_download_params(model_path):
+    if os.path.exists(model_path) or not is_link(model_path):
+        return model_path
+    else:
+        url = model_path
+    tmp_path = os.path.join(MODELS_DIR, url.split("/")[-1])
+    print("download {} to {}".format(url, tmp_path))
+    os.makedirs(MODELS_DIR, exist_ok=True)
+    download_with_progressbar(url, tmp_path)
+    return tmp_path
+
+
 def is_link(s):
-    return s is not None and s.startswith('http')
+    return s is not None and s.startswith("http")
 
 
 def confirm_model_dir_url(model_dir, default_model_dir, default_url):
@@ -76,7 +89,7 @@ def confirm_model_dir_url(model_dir, default_model_dir, default_url):
     if model_dir is None or is_link(model_dir):
         if is_link(model_dir):
             url = model_dir
-        file_name = url.split('/')[-1][:-4]
+        file_name = url.split("/")[-1][:-4]
         model_dir = default_model_dir
         model_dir = os.path.join(model_dir, file_name)
     return model_dir, url
